Refactor DID entities to another directory
The DID entities are refactored to another location
in order to verifiable credentials entities and more
functionality clearer.
