+++ src/Verifiable.Core/Credential/VerifiableCredentials.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace Verifiable.Core
 {
     //https://www.w3.org/TR/vc-data-model/#types
