+++ src/Verifiable.Core/Did/KeyFormat.cs
@@ -1,0 +1,7 @@
+namespace Verifiable.Core.Did
+{
+    /// <summary>
+    /// https://www.w3.org/TR/did-core/#key-types-and-formats
+    /// </summary>
+    public abstract class KeyFormat { }
+}
