+++ src/Verifiable.Core/ICoder.cs
@@ -1,5 +1,4 @@
 using SimpleBase;
-using System;
 
 namespace Verifiable.Core
 {
